requirements.txt
api==0.0.7
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
djangorestframework==3.7.7
idna==2.6
nltk==3.2.5
nose==1.3.7
numpy==1.14.1
pytz==2018.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.2
task==0.1
textblob==0.15.1
tweet-preprocessor==0.5.0
urllib3==1.22
